+++ infrastructure-android/src/main/kotlin/android/AndroidApplicationPlugin.kt
@@ -11,10 +11,7 @@ import com.redmadrobot.build.android.internal.android
 import com.redmadrobot.build.android.internal.androidComponents
 import com.redmadrobot.build.android.internal.projectProguardFiles
 import com.redmadrobot.build.android.task.MakeDebuggableTask
-import com.redmadrobot.build.dsl.BUILD_TYPE_DEBUG
-import com.redmadrobot.build.dsl.BUILD_TYPE_QA
-import com.redmadrobot.build.dsl.BUILD_TYPE_RELEASE
-import com.redmadrobot.build.dsl.finalizeQaBuildType
+import com.redmadrobot.build.dsl.*
 import com.redmadrobot.build.internal.InternalGradleInfrastructureApi
 import org.gradle.api.Project
 import org.gradle.api.tasks.TaskContainer
@@ -44,9 +41,6 @@ private fun Project.configureApp() = android<ApplicationExtension> {
     defaultConfig {
         // Collect proguard rules from 'proguard' dir
         setProguardFiles(projectProguardFiles() + getDefaultProguardFile("proguard-android-optimize.txt"))
-
-        buildConfigField("boolean", VAR_LOCK_ORIENTATION, "true")
-        buildConfigField("boolean", VAR_CRASH_REPORTS_ENABLED, "true")
     }
 
     finalizeQaBuildType()
@@ -56,9 +50,6 @@ private fun Project.configureApp() = android<ApplicationExtension> {
             isDebuggable = true
             isMinifyEnabled = false
             isShrinkResources = false
-
-            buildConfigField("boolean", VAR_LOCK_ORIENTATION, "false")
-            buildConfigField("boolean", VAR_CRASH_REPORTS_ENABLED, "false")
         }
 
         release {
@@ -73,10 +64,6 @@ private fun Project.configureApp() = android<ApplicationExtension> {
             applicationIdSuffix = ".$BUILD_TYPE_QA"
             matchingFallbacks += listOf(BUILD_TYPE_DEBUG, BUILD_TYPE_RELEASE)
             signingConfig = signingConfigs.findByName(BUILD_TYPE_DEBUG)
-
-            // We can not use isDebuggable = true here, so set DEBUG field ourselves.
-            // See `makeDebuggable` for more information
-            buildConfigField(type = "boolean", name = "DEBUG", value = "true")
         }
     }
 }
@@ -108,9 +95,14 @@ private fun ApplicationExtension.finalizeApp(
         xmlOutput = staticAnalyzerSpec.reportsDir.file("lint-results.xml").get().asFile
         htmlOutput = staticAnalyzerSpec.reportsDir.file("lint-results.html").get().asFile
     }
+
+    buildTypes {
+        if (buildFeatures.buildConfig == true) {
+            qa {
+                // We can not use isDebuggable = true here, so set DEBUG field ourselves.
+                // See `makeDebuggable` for more information
+                buildConfigField(type = "boolean", name = "DEBUG", value = "true")
+            }
+        }
+    }
 }
-
-// Constants
-
-private const val VAR_LOCK_ORIENTATION = "LOCK_ORIENTATION"
-private const val VAR_CRASH_REPORTS_ENABLED = "CRASH_REPORTS_ENABLED"