+++ CHANGELOG.md
@@ -22,10 +22,12 @@ android {
   redmadrobot.add.repositories=true
   ```
 - **android:** Default `targetSdk` changed from `33` to `34`
+- **android:** Do not add `LOCK_ORIENTATION` and `CRASH_REPORTS_ENABLED` variables to `BuildConfig` implicitly
 
 ### Other Changes
 
 - **android:** Remove the workaround for Explicit API enabling as the issue has been [fixed](https://youtrack.jetbrains.com/issue/KT-37652) in Kotlin 1.9
+- **android:** Remove disabling of build features `aidl`, `renderScript` and `buildConfig` as they are already disabled by default in new versions of AGP
 - **kotlin:** Deprecate accessor `kotlinCompile`.
   It is recommended to use `kotlin.compilerOptions { ... }` to configure compilation instead.
 - Update Gradle to `8.8`